Encapsulate DOM-specific theming code in ThemeManager class
