+++ src/domain/session/settings/SettingsViewModel.js
@@ -16,7 +16,6 @@ limitations under the License.
 
 import {ViewModel} from "../../ViewModel.js";
 import {SessionBackupViewModel} from "./SessionBackupViewModel.js";
-import {getStyleSheetElements, setTheme} from "../../../platform/web/ui/general/theming.js";
 
 class PushNotificationStatus {
     constructor() {
@@ -47,7 +46,6 @@ export class SettingsViewModel extends ViewModel {
         this._sessionBackupViewModel = this.track(new SessionBackupViewModel(this.childOptions({session})));
         this._closeUrl = this.urlCreator.urlUntilSegment("session");
         this._estimate = null;
-        this._styleSheetElements = getStyleSheetElements();
 
         this.sentImageSizeLimit = null;
         this.minSentImageSizeLimit = 400;
@@ -162,21 +160,17 @@ export class SettingsViewModel extends ViewModel {
     }
 
     get themes() {
-        return Object.keys(this._styleSheetElements);
+        return this.platform.themeManager.themes;
     }
 
     get currentTheme() {
-        for (const name in this._styleSheetElements) {
-            if (!this._styleSheetElements[name].disabled) {
-                return name;
-            }
-        }
+        return this.platform.themeManager.currentTheme;
     }
 
     async setTheme(name) {
         const vm = this;
         return new Promise(resolve => {
-            setTheme(name, this._styleSheetElements).then(() => {
+            vm.platform.themeManager.setTheme(name).then(() => {
                 vm.platform.settingsStorage.setString("theme", name);
                 vm.emitChange("currentTheme");
                 resolve();
